wizard, uplink-files: add uci-setting to request different auth-types
add uci "ffberlin-uplink.uplink.auth"to give each uplink-provider
a way to have the assistent asking for different kinds of
authentication. Following types are supported:
- none: the uplink-type don't need additional parameters
- x509: the uplink-type requires a set of X.509 certificate
        (stored in /etc/openvpn/, as for VPN03 and tunnel-berlin)
- pass: todo - just ask for a simple password / psk
- userpass: todo - ask for username and password

+++ utils/luci-app-ffwizard-berlin/luasrc/model/cbi/freifunk/assistent/shareInternet.lua
@@ -10,27 +10,28 @@ f.reset = false
 css = f:field(DummyValue, "css", "")
 css.template = "freifunk/assistent/snippets/css"
 
-vpninfo = f:field(DummyValue, "vpninfo", "")
-vpninfo.template = "freifunk/assistent/snippets/vpninfo"
+if uci.get("ffberlin-uplink", "uplink", "auth") == "x509" then
+  vpninfo = f:field(DummyValue, "vpninfo", "")
+  vpninfo.template = "freifunk/assistent/snippets/vpninfo"
+  if luci.http.formvalue("reupload", true) == "1" then
+    fs.unlink("/etc/openvpn/ffuplink.crt")
+    fs.unlink("/etc/openvpn/ffuplink.key")
+    luci.http.redirect(luci.dispatcher.build_url("admin/freifunk/assistent/sharedInternet"))
+  end
+  if fs.access("/etc/openvpn/ffuplink.crt") then
+    vpncertreupload = f:field(DummyValue, "reupload", "")
+    vpncertreupload.template = "freifunk/assistent/snippets/vpncertreupload"
+  else
+    local cert = f:field(FileUpload, "cert", translate("Local Certificate"),"ffuplink.crt")
+    cert.default="/etc/openvpn/ffuplink.crt"
+    cert.rmempty = false
+    cert.optional = false
 
-if luci.http.formvalue("reupload", true) == "1" then
-  fs.unlink("/etc/openvpn/ffuplink.crt")
-  fs.unlink("/etc/openvpn/ffuplink.key")
-  luci.http.redirect(luci.dispatcher.build_url("admin/freifunk/assistent/sharedInternet"))
-end
-if fs.access("/etc/openvpn/ffuplink.crt") then
-  vpncertreupload = f:field(DummyValue, "reupload", "")
-  vpncertreupload.template = "freifunk/assistent/snippets/vpncertreupload"
-else
-  local cert = f:field(FileUpload, "cert", translate("Local Certificate"),"ffuplink.crt")
-  cert.default="/etc/openvpn/ffuplink.crt"
-  cert.rmempty = false
-  cert.optional = false
-
-  local key = f:field(FileUpload, "key", translate("Local Key"),"ffuplink.key")
-  key.default="/etc/openvpn/ffuplink.key"
-  key.rmempty = false
-  key.optional = false
+    local key = f:field(FileUpload, "key", translate("Local Key"),"ffuplink.key")
+    key.default="/etc/openvpn/ffuplink.key"
+    key.rmempty = false
+    key.optional = false
+  end
 end
 
 shareBandwidth = f:field(DummyValue, "shareBandwidthfo", "")
